+++ crates/goose-mcp/src/developer/analyze/languages/ruby.rs
@@ -113,7 +113,7 @@ fn find_first_method_in_class(
     source: &str,
     max_depth: usize,
 ) -> Option<String> {
-    for i in 0..class_node.child_count() {
+    for i in 0..class_node.child_count() as u32 {
         if let Some(child) = class_node.child(i) {
             if child.kind() == "body_statement" {
                 return find_method_in_body_with_depth(&child, source, 0, max_depth);
@@ -134,10 +134,10 @@ fn find_method_in_body_with_depth(
         return None;
     }
 
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             if child.kind() == "method" {
-                for j in 0..child.child_count() {
+                for j in 0..child.child_count() as u32 {
                     if let Some(name_node) = child.child(j) {
                         if name_node.kind() == "identifier" {
                             return source.get(name_node.byte_range()).map(|s| s.to_string());
